refactor: add src for stance.hpp
